+++ miden-tx/src/tests/kernel_tests/test_prologue.rs
@@ -16,7 +16,6 @@ use miden_lib::transaction::{
     TransactionKernel,
 };
 use miden_objects::{
-    assembly::ProgramAst,
     testing::{
         account::AccountBuilder,
         constants::FUNGIBLE_FAUCET_INITIAL_BALANCE,
@@ -51,20 +50,18 @@ fn test_transaction_prologue() {
         end
         ";
 
-    let mock_tx_script_code = ProgramAst::parse(
-        "
+    let mock_tx_script_code = "
         begin
             push.1.2.3.4 dropw
         end
-        ",
-    )
-    .unwrap();
-    let (tx_script, _) = TransactionScript::new(
-        mock_tx_script_code,
-        vec![],
-        &TransactionKernel::assembler().with_debug_mode(true),
-    )
-    .unwrap();
+        ";
+
+    let mock_tx_script_program = TransactionKernel::assembler()
+        .with_debug_mode(true)
+        .assemble_program(mock_tx_script_code)
+        .unwrap();
+
+    let tx_script = TransactionScript::new(mock_tx_script_program, vec![]);
 
     let note_args = [
         [Felt::new(91), Felt::new(91), Felt::new(91), Felt::new(91)],
@@ -125,7 +122,7 @@ fn global_input_memory_assertions(process: &Process<MockHost>, inputs: &Transact
 
     assert_eq!(
         read_root_mem_value(process, TX_SCRIPT_ROOT_PTR),
-        **inputs.tx_args().tx_script().as_ref().unwrap().hash(),
+        *inputs.tx_args().tx_script().as_ref().unwrap().hash(),
         "The transaction script root should be stored at the TX_SCRIPT_ROOT_PTR"
     );
 }